requirements.txt
ccxt
pandas
numpy
ta
scikit-learn
xgboost
python-dotenv
requests
